PIC, PIT refactor
